test(research): extract fallback human supersession
